ci: pull cargo test out of the build CI job
Building the ci-cargo-test Docker image is one of the slowest parts of
the CI build job because it requires shipping around a 1.5GiB image
that's only used by one other job.

Instead, use a third builder agent to run `cargo nextest` directly. This
is a bit more wasteful in terms of overall CPU usage (we have to
recompile the world from scratch twice), but should make CI *much*
faster because the main build job will complete much faster, which lets
the rest of the tests get started much sooner.

Fix #13010.